bug fixes for project1

+++ project1/CMYKtoRGB.java
@@ -11,8 +11,8 @@ public class CMYKtoRGB {
 		double green = 255 * white * (1 - magenta);
 		double blue = 255 * white * (1 - yellow);
 
-		System.out.println("red   = " + red);
-		System.out.println("green = " + green);
-		System.out.println("blue  = " + blue);
+		System.out.println("red   = " + (int) red);
+		System.out.println("green = " + (int) green);
+		System.out.println("blue  = " + (int) blue);
 	}
 }
